Implement perf tests for FCMP++ verification
