+++ tests/CMakeLists.txt
@@ -58,8 +58,8 @@ else ()
   endif()
   add_subdirectory(crypto)
   add_subdirectory(functional_tests)
-  add_subdirectory(performance_tests)
   add_subdirectory(unit_tests)
+  add_subdirectory(performance_tests)
   add_subdirectory(difficulty)
   add_subdirectory(block_weight)
   add_subdirectory(hash)
