+++ tests/unit_tests/CMakeLists.txt
@@ -103,6 +103,7 @@ set(unit_tests_sources
   tx_verification_utils.cpp
   hardfork.cpp
   unbound.cpp
+  unit_tests_utils.cpp
   uri.cpp
   util.cpp
   varint.cpp
@@ -160,6 +161,10 @@ set_property(TARGET unit_tests
   APPEND_STRING
   PROPERTY
     COMPILE_FLAGS " -Wno-undef -Wno-sign-compare")
+
+add_library(unit_tests_utils
+  STATIC
+    unit_tests_utils.cpp)
 
 SET_PROPERTY(SOURCE memwipe.cpp PROPERTY COMPILE_FLAGS -Ofast)
