+++ tests/unit_tests/fcmp_pp.cpp
@@ -317,81 +317,6 @@ static std::string get_fcmp_pp_filename(const std::size_t n_inputs)
     return unit_test::data_dir.string() + "/fcmp_pp_verify_inputs_" + std::to_string(n_inputs) + "in.bin";
 }
 //----------------------------------------------------------------------------------------------------------------------
-/*
-static void write_fcmp_pp_verify_input_to_file(const std::size_t n_inputs,
-    const crypto::hash &signable_tx_hash,
-    const fcmp_pp::FcmpPpProof &fcmp_pp_proof,
-    const uint8_t n_layers,
-    const crypto::ec_point &tree_root_bytes,
-    const std::vector<crypto::ec_point> &pseudo_outs,
-    const std::vector<crypto::key_image> &key_images)
-{
-    const std::string filename = get_fcmp_pp_filename(n_inputs);
-    LOG_PRINT_L1("Writing FCMP++ proof to " << filename);
-    std::ofstream file(filename, std::ios::binary);
-    ASSERT_FALSE(!file);
-
-    file.write(reinterpret_cast<const char*>(&signable_tx_hash), sizeof(signable_tx_hash));
-
-    file.write(reinterpret_cast<const char*>(&n_layers), sizeof(uint8_t));
-    file.write(reinterpret_cast<const char*>(fcmp_pp_proof.data()), fcmp_pp_proof.size());
-
-    file.write(reinterpret_cast<const char*>(&tree_root_bytes), sizeof(tree_root_bytes));
-
-    CHECK_AND_ASSERT_THROW_MES(pseudo_outs.size() == n_inputs, "Unexpected size pseudo outs");
-    for (const auto &po : pseudo_outs)
-        file.write(reinterpret_cast<const char*>(&po), sizeof(po));
-
-    CHECK_AND_ASSERT_THROW_MES(key_images.size() == n_inputs, "Unexpected size key_images");
-    for (const auto &ki : key_images)
-        file.write(reinterpret_cast<const char*>(&ki), sizeof(ki));
-
-    file.close();
-}
-*/
-//----------------------------------------------------------------------------------------------------------------------
-static void read_fcmp_pp_verify_input_from_file(const std::size_t n_inputs,
-    crypto::hash &signable_tx_hash,
-    fcmp_pp::FcmpPpProof &fcmp_pp_proof,
-    uint8_t &n_layers,
-    fcmp_pp::TreeRootShared &tree_root,
-    std::vector<crypto::ec_point> &pseudo_outs,
-    std::vector<crypto::key_image> &key_images)
-{
-    const std::string filename = get_fcmp_pp_filename(n_inputs);
-    std::ifstream file(filename, std::ios::binary);
-    ASSERT_FALSE(!file);
-
-    file.read(reinterpret_cast<char*>(&signable_tx_hash), sizeof(signable_tx_hash));
-
-    file.read(reinterpret_cast<char*>(&n_layers), sizeof(n_layers));
-    const std::size_t proof_len = fcmp_pp::fcmp_pp_proof_len(n_inputs, n_layers);
-    fcmp_pp_proof.resize(proof_len);
-    file.read(reinterpret_cast<char*>(fcmp_pp_proof.data()), proof_len);
-
-    const auto curve_trees = fcmp_pp::curve_trees::curve_trees_v1();
-    crypto::ec_point tree_root_bytes;
-    file.read(reinterpret_cast<char*>(&tree_root_bytes), sizeof(tree_root_bytes));
-    tree_root = n_layers % 2 == 0
-        ? fcmp_pp::helios_tree_root(curve_trees->m_c2->from_bytes(tree_root_bytes))
-        : fcmp_pp::selene_tree_root(curve_trees->m_c1->from_bytes(tree_root_bytes));
-
-    pseudo_outs = std::vector<crypto::ec_point>(n_inputs);
-    for (auto &po : pseudo_outs)
-        file.read(reinterpret_cast<char*>(&po), sizeof(po));
-
-    key_images = std::vector<crypto::key_image>(n_inputs);
-    for (auto &ki : key_images)
-        file.read(reinterpret_cast<char*>(&ki), sizeof(ki));
-
-    file.close();
-
-    MDEBUG("signable_tx_hash: " << signable_tx_hash <<
-        ", proof_size: "              << proof_len <<
-        ", n_layers: "                << (std::size_t)n_layers <<
-        ", tree_root_bytes: "         << tree_root_bytes);
-}
-//----------------------------------------------------------------------------------------------------------------------
 //----------------------------------------------------------------------------------------------------------------------
 TEST(fcmp_pp, prove)
 {
@@ -721,15 +646,14 @@ TEST(fcmp_pp, verify)
             membership_proof,
             n_layers);
 
-/*
-        write_fcmp_pp_verify_input_to_file(n_inputs,
-            signable_tx_hash,
-            fcmp_pp_proof,
-            n_layers,
-            tree_root_bytes,
-            pseudo_outs,
-            key_images);
-*/
+        // unit_test::write_fcmp_pp_verify_input_to_file(get_fcmp_pp_filename(n_inputs),
+        //     n_inputs,
+        //     signable_tx_hash,
+        //     fcmp_pp_proof,
+        //     n_layers,
+        //     tree_root_bytes,
+        //     pseudo_outs,
+        //     key_images);
 
         LOG_PRINT_L1("Verifying (n_inputs=" << n_inputs << ")");
         bool verify = fcmp_pp::verify(
@@ -757,7 +681,8 @@ TEST(fcmp_pp, batch_verify_from_file)
     std::vector<crypto::ec_point> pseudo_outs;
     std::vector<crypto::key_image> key_images;
 
-    read_fcmp_pp_verify_input_from_file(n_inputs,
+    unit_test::read_fcmp_pp_verify_input_from_file(get_fcmp_pp_filename(n_inputs),
+        n_inputs,
         signable_tx_hash,
         fcmp_pp_proof,
         n_layers,
